Result: show product's latest price
